% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{canton_zh_no_lakes}
\alias{canton_zh_no_lakes}
\title{Grenzen von ZH}
\format{
sfc_MULTIPOLYGON
}
\description{
Geographische Grenzen der Kanton ZH ohne Seen (Quelle: OpenStreetMap).
}
\keyword{datasets}
