% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{Farben}
\alias{Farben}
\alias{Akzentfarben}
\alias{AkzentfarbenDarkmode}
\alias{AkzentfarbenSoft}
\alias{Funktion}
\alias{Grautoene}
\alias{Infografiken}
\alias{ktz_palette}
\title{Color palettes for Kanton Zürich}
\format{
A named vector of color codes
}
\description{
Color palettes conform with the standards of Kanton Zürich.\cr
Palette names:\cr
\itemize{
\item Akzentfarben
\item AkzentfarbenDarkmode
\item AkzentfarbenSoft
\item Funktion
\item Grautoene
\item Infografiken
\item ktz_palette
}
}
\keyword{datasets}
