% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset_utils.R
\name{lighting_condition}
\alias{lighting_condition}
\title{Lighting Condition}
\usage{
lighting_condition(lat, lng, time)
}
\arguments{
\item{lat, long}{Latitude and longitude in degrees (can be vectors).}

\item{time}{A datetime or a vector of datetimes.}
}
\description{
What part of the day is it? Given lat, lng and time you get what part of the
day it is (day, night, civil, nautical or astronomical twilight).
This function does not work if there is no astronomical twilight during that
day in that place, see
https://en.wikipedia.org/wiki/Twilight#Between_day_and_night
}
