% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{old_df}
\alias{old_df}
\title{Historische Daten}
\format{
dataframe
}
\description{
Historische Unfall- und Fahzeugbestand-daten der Schweiz, nach Kanton, Schwere und Jahr (1975-2023).\cr
Quelle (BFS):
\itemize{
\item https://www.bfs.admin.ch/asset/de/31726346
\item https://www.bfs.admin.ch/asset/de/33827606
\item https://www.bfs.admin.ch/asset/de/33827620
}
}
\keyword{datasets}
