% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SafetyCockpit-package.R
\docType{package}
\name{SafetyCockpit-package}
\alias{SafetyCockpit-package}
\title{SafetyCockpit: Safety Cockpit}
\description{
A Cockpit to visualize the ongoing accidents in Kanton Zürich.
}
\author{
\strong{Maintainer}: FaSi \email{sicherheit.tba@bd.zh.ch}

Authors:
\itemize{
  \item Enea Rossi \email{enea.rossi.8@gmail.com}
  \item Daniele Valerio \email{daniele.valerio@zh.ch}
}

}
\keyword{internal}
