% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_cockpit.R
\name{SafetyCockpit}
\alias{SafetyCockpit}
\title{Safety Cockpit}
\usage{
SafetyCockpit(..., load_data = T, skip_login = F)
}
\arguments{
\item{...}{Arguments passed on to shiny::ShinyApp()}

\item{load_data}{Whether to load the data from the SQL Server (useful if the data is already loaded).}

\item{skip_login}{Whether to skip the log in page (useful while working locally).}
}
\value{
An object that represents the app.
}
\description{
Launches the Safety cockpit
}
