% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset_utils.R
\name{sunriset}
\alias{sunriset}
\title{Sunrise and Sunset}
\usage{
sunriset(
  lat,
  lng,
  date,
  extra_cols = F,
  tz_out = NULL,
  h_o_deg = -0.8333,
  prefix = ""
)
}
\arguments{
\item{lat, long}{Latitude and longitude in degrees (can be vectors).}

\item{date}{The date to be considered (can be vector).}

\item{extra_cols}{Should extra columns be returned?}

\item{tz_out}{By default the function tries to infer the output timezone from
the system, this is an override for it. Must be a string which as.POSIXct can handle or NULL.}

\item{h_o_deg}{Angle to the horizon in degrees. The default value is generally
correct, but can be adjusted for minor correction or changed completely to get
civil (-6°), nautical (-12°) and astronomical (-18°) twilight.}
}
\description{
Calculates the rise and set of the sun given coordinates and a date.
By default it calculates the regular rise and set times, but one can adjust
h_o_deg to get civil, nautical and astronomical twilight as well.
Based on https://en.wikipedia.org/wiki/Sunrise_equation#Complete_calculation_on_Earth
}
