% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset_utils.R
\name{sunriset_df}
\alias{sunriset_df}
\title{Sunriset for dataframes}
\usage{
sunriset_df(df, join = T, ...)
}
\arguments{
\item{df}{A dataframe with columns lat, lng and datum.}

\item{join}{If True then the function returns a column bind of the original
dataframe with the computed columns (will create duplicates if extra_cols = True).}

\item{...}{Other Arguments passed on to sunriset}
}
\description{
Method to apply the function sunriset to a dataframe (or tibble) with some
predefined columns.
}
